// frontend/src/api/generated/main-service/apiGenerated.ts
export interface FontDetailDto {
  /** @format uuid */
  id: string
  name: string
  familyName: string
  subFamilyName: string
  /**
   * @format int16
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  weight: number | string
  isItalic: boolean
  fileFormat: string
  /**
   * @format int64
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  fileSizeBytes: number | string
  isSystemFont: boolean
  accessType: string
  /** @format date-time */
  createdDatetime: any
  /** @format uuid */
  ownershipId?: null | string
  licenseNote?: null | string
  downloadUrl?: null | string
}

export interface FontSummaryDto {
  /** @format uuid */
  id: string
  name: string
  familyName: string
  subFamilyName: string
  /**
   * @format int16
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  weight: number | string
  isItalic: boolean
  fileFormat: string
  /**
   * @format int64
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  fileSizeBytes: number | string
  isSystemFont: boolean
  accessType: string
  /** @format date-time */
  createdDatetime: any
}